requirements.txt
Django==5.1.7
djangorestframework==3.15.2
environ==1.0
Requests==2.32.3
